% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMM.R
\name{GLMM}
\alias{GLMM}
\title{GLMM package with CTSIB.csv}
\usage{
GLMM(data = "ctsib.csv")
}
\arguments{
\item{data}{the data set for my example}
}
\value{
A list with summary statistics from the fitted model
}
\description{
GLMM package with CTSIB.csv
}
\examples{
model <- GLMM("ctsib.csv")
GLMM function can create a GLMM model, whose summary statistics are returned.
}
