% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMM.R
\name{GLMM}
\alias{GLMM}
\title{Generalized linear mixed effects model function package with data set CTSIB.csv}
\usage{
GLMM(data)
}
\arguments{
\item{data}{the data set for analysis}
}
\value{
A list with summary statistics from the fitted model including fixed effect, random variance, and random effect.
}
\description{
Generalized linear mixed effects model function package with data set CTSIB.csv
}
\examples{
model <- GLMM("ctsib.csv")
the parameter should be a csv data set file
GLMM function can create a GLMM model, that will return summary of statistics after analysis.
}
\author{
Yichao Liu, Weixuan Zhang
}
